Update BicycleStringsBundle_fr_FR.properties
I am not sure if there are abbreviations for the days of the week in french as there are in English but we may want them otherwise the calender may look stretched.  Also we may need the months translated to make it easier for the user to read the date.