Added text box with + on it for country code and more
Added it to the user/worker view(where we insert new ones) but no others
just because we dont have time to parse the phone number that we get
from db to put into country code, but i changed the messages from
warnings to error's with red x on all the views in UI. I think I checked
and made sure the title is also from the localized bundle, but double
check for me. I changed the add employee button and label at top of that
view to newWorker key which says insert new employee on button and
label. It just has the !!!!!!! at the end of it so make sure to edit
that out if you arent changing it.